Порт на AVR/Arduino, значительные переработки в области используемых типов данных

+++ stepper_h/examples/draw_triangle/draw_triangle.ino
@@ -13,18 +13,18 @@ static void prepare_line1() {
     // line1: (0,0,0) -> (150000000,50000000,20000000)
     
     // X - длинная координата - ее проходим с максимальной скоростью
-    //int steps_x = 150000000 / 7500;
-    int steps_x = (150000000 - sm_x.current_pos) / 7500;
-    int delay_x = 1000;
-    int time_x = delay_x * abs(steps_x);
+    //long steps_x = 150000000 / 7500;
+    long steps_x = (150000000 - sm_x.current_pos) / 7500;
+    unsigned long delay_x = 1000;
+    unsigned long time_x = delay_x * abs(steps_x);
     
-    //int steps_y = 50000000 / 7500;
-    int steps_y = (50000000 - sm_y.current_pos) / 7500;
-    int delay_y = time_x / abs(steps_y);
+    //long steps_y = 50000000 / 7500;
+    long steps_y = (50000000 - sm_y.current_pos) / 7500;
+    unsigned long delay_y = time_x / abs(steps_y);
     
-    //int steps_z = 20000000 / 7500;
-    int steps_z = (20000000 - sm_z.current_pos) / 7500;
-    int delay_z = time_x / abs(steps_z);
+    //long steps_z = 20000000 / 7500;
+    long steps_z = (20000000 - sm_z.current_pos) / 7500;
+    unsigned long delay_z = time_x / abs(steps_z);
 
     // prepare_steps(stepper *smotor,
     //     long step_count, unsigned long step_delay,
@@ -48,19 +48,19 @@ static void prepare_line2() {
     // реально получится:
     // steps_x=-13333, steps_y=13334
     
-    //int steps_y = (150000000 - 50000000) / 7500;
-    int steps_y = (150000000 - sm_y.current_pos) / 7500;
-    int delay_y = 1000;
-    int time_y = delay_y * abs(steps_y);
+    //long steps_y = (150000000 - 50000000) / 7500;
+    long steps_y = (150000000 - sm_y.current_pos) / 7500;
+    unsigned long delay_y = 1000;
+    unsigned long time_y = delay_y * abs(steps_y);
     
-    //int steps_x = (50000000 - 150000000) / 7500;
-    int steps_x = (50000000 - sm_x.current_pos) / 7500;
-    int delay_x = time_y / abs(steps_y);
+    //long steps_x = (50000000 - 150000000) / 7500;
+    long steps_x = (50000000 - sm_x.current_pos) / 7500;
+    unsigned long delay_x = time_y / abs(steps_y);
     
     // путь по z=0
-    //int steps_z = (20000000 - 20000000) / 7500;
-    //int steps_z = (20000000 - sm_z.current_pos) / 7500;
-    //int delay_z = time_x / abs(steps_z);
+    //long steps_z = (20000000 - 20000000) / 7500;
+    //long steps_z = (20000000 - sm_z.current_pos) / 7500;
+    //unsigned long delay_z = time_x / abs(steps_z);
 
     // prepare_steps(stepper *smotor,
     //     long step_count, unsigned long step_delay,
@@ -80,18 +80,18 @@ static void prepare_line3() {
     // line3: (50000000,150000000,20000000) -> (0,0,0)
     
     // Y - длинная координата - ее проходим с максимальной скоростью
-    //int steps_y = -150000000 / 7500;
-    int steps_y = (0 - sm_y.current_pos) / 7500;
-    int delay_y = 1000;
-    int time_y = delay_y * abs(steps_y);
+    //long steps_y = -150000000 / 7500;
+    long steps_y = (0 - sm_y.current_pos) / 7500;
+    unsigned long delay_y = 1000;
+    unsigned long time_y = delay_y * abs(steps_y);
     
-    //int steps_x = -50000000 / 7500;
-    int steps_x = (0 - sm_x.current_pos) / 7500;
-    int delay_x = time_y / abs(steps_x);
+    //long steps_x = -50000000 / 7500;
+    long steps_x = (0 - sm_x.current_pos) / 7500;
+    unsigned long delay_x = time_y / abs(steps_x);
     
-    //int steps_z = -20000000 / 7500;
-    int steps_z = (0 - sm_z.current_pos) / 7500;
-    int delay_z = time_y / abs(steps_z);
+    //long steps_z = -20000000 / 7500;
+    long steps_z = (0 - sm_z.current_pos) / 7500;
+    unsigned long delay_z = time_y / abs(steps_z);
 
     // prepare_steps(stepper *smotor,
     //     long step_count, unsigned long step_delay,
@@ -108,8 +108,8 @@ void setup() {
     // connected stepper motors
     // init_stepper(stepper* smotor, char name,
     //     int pin_step, int pin_dir, int pin_en,
-    //     bool invert_dir, int step_delay,
-    //     int distance_per_step)
+    //     bool invert_dir, unsigned long step_delay,
+    //     unsigned long distance_per_step)
     // init_stepper_ends(stepper* smotor,
     //     end_strategy min_end_strategy, end_strategy max_end_strategy,
     //     long long min_pos, long long max_pos);
@@ -163,10 +163,10 @@ void loop() {
         stepper_start_cycle();
     }
 
-    static int prevTime = 0;
+    static unsigned long prevTime = 0;
     // Debug messages - print current positions of motors once per second
     // while they are rotating, once per 10 seconds when they are stopped
-    int currTime = millis();
+    unsigned long currTime = millis();
     if( (stepper_cycle_running() && (currTime - prevTime) >= 1000) || (currTime - prevTime) >= 10000 ) {
         prevTime = currTime;
         Serial.print("X.pos=");
